[FSDE-31720] Update Cypress version in Chi to latest version
