+++ cicd/jenkins/Jenkinsfile
@@ -473,9 +473,8 @@ pipeline {
           }
           stage ('Cypress local Testing') {
             agent {
-              dockerfile {
-                filename 'Dockerfile'
-                dir 'cicd/docker/sonar-scanner'
+              docker {
+                image "nexusprod.corp.intranet:4567/jsl-docker-standard-images/node14_agent:latest"
                 label 'Docker-enabled'
                 args '-m 2G --shm-size=2gb'
               }
