+++ cypress/e2e/chi-custom-elements/drawer.cy.js
@@ -1,7 +1,6 @@
 /// <reference types="Cypress" />
 
 describe('Drawer', function() {
-
   it('Drawer custom element should initiate a Chi drawer component. ', function() {
     cy.visit('tests/custom-elements/drawer.html');
 
@@ -15,7 +14,6 @@ describe('Drawer', function() {
   });
 
   it('By default drawer custom element should include a close button. ', function() {
-
     cy.get('[data-cy="test-top-headed"]')
       .should('have.class', 'hydrated')
       .children()
@@ -25,11 +23,9 @@ describe('Drawer', function() {
       .should('have.length', 1)
       .find('button.chi-button chi-icon[icon="x"]')
       .should('exist');
-
   });
 
   it('Should not include a close button when non-closable="" attribute is present ', function() {
-
     cy.get('[data-cy="test-non-closable"]')
       .should('have.class', 'hydrated')
       .children()
@@ -37,18 +33,19 @@ describe('Drawer', function() {
       .should('match', 'div.chi-drawer.-active')
       .find('button.chi-button chi-icon[icon="x"]')
       .should('not.exist');
-
   });
 
   it('Should show a backdrop if configured to do so. ', function() {
-
     cy.get('[data-cy="test-backdrop-top"]')
       .should('have.class', 'hydrated')
-      .children().first()
+      .children()
+      .first()
       .should('match', 'div.chi-backdrop.-animated')
-      .children().first()
+      .children()
+      .first()
       .should('match', 'div.chi-backdrop__wrapper')
-      .children().first()
+      .children()
+      .first()
       .should('match', 'div.chi-drawer.-top.-active');
   });
 
@@ -64,7 +61,6 @@ describe('Drawer', function() {
   });
 
   it('Should close when clicking the close button. ', function() {
-
     cy.get('[data-cy="test-top-headed"]')
       .children()
       .first()
@@ -73,11 +69,9 @@ describe('Drawer', function() {
       .get('[data-cy="test-top-headed"]')
       .should('not.match', '[active]')
       .should('not.be.visible');
-
   });
 
   it('Should have a title if attribute title="" is provided ', function() {
-
     cy.get('[data-cy="test-top-with-title"]')
       .children()
       .first()
@@ -85,13 +79,11 @@ describe('Drawer', function() {
       .find('.chi-drawer__title')
       .contains('Drawer title here')
       .should('exist');
-
   });
 
   it('Should close when calling the hide method. ', function() {
-
     cy.get('[data-cy="test-backdrop-top-headed"]')
-      .then(function(drawer){
+      .then(function(drawer) {
         drawer[0].hide();
         return new Promise(resolve => resolve(drawer));
       })
@@ -102,9 +94,8 @@ describe('Drawer', function() {
   });
 
   it('Should show when calling the show method. ', function() {
-
     cy.get('[data-cy="test-backdrop-top-headed"]')
-      .then(function(drawer){
+      .then(function(drawer) {
         drawer[0].show();
         return new Promise(resolve => resolve(drawer));
       })
@@ -112,13 +103,11 @@ describe('Drawer', function() {
       .should('match', '[active]')
       .find('.chi-drawer')
       .should('be.visible');
-
   });
 
   it('Should show when calling toggle method and is hidden and vice versa. ', function() {
-
     cy.get('[data-cy="test-backdrop-right-headed"]')
-      .then(function(drawer){
+      .then(function(drawer) {
         drawer[0].toggle();
         return new Promise(resolve => resolve(drawer));
       })
@@ -127,7 +116,7 @@ describe('Drawer', function() {
       .find('.chi-drawer')
       .should('not.be.visible')
       .get('[data-cy="test-backdrop-right-headed"]')
-      .then(function(drawer){
+      .then(function(drawer) {
         drawer[0].toggle();
         return new Promise(resolve => resolve(drawer));
       })
@@ -135,6 +124,5 @@ describe('Drawer', function() {
       .should('match', '[active]')
       .find('.chi-drawer')
       .should('be.visible');
-
   });
 });
