+++ cypress/e2e/chi-custom-elements/progress.cy.js
