+++ cypress/e2e/chi-custom-elements/search-input.cy.js
@@ -1,6 +1,6 @@
 const SEARCH_INPUT_CY = {
-  AUTOCOMPLETE: '[data-cy="autocomplete"]',
-}
+  AUTOCOMPLETE: '[data-cy="autocomplete"]'
+};
 
 describe('Search input', () => {
   before(() => {
@@ -10,11 +10,14 @@ describe('Search input', () => {
   describe('Autocomplete', () => {
     beforeEach(() => {
       cy.get(SEARCH_INPUT_CY.AUTOCOMPLETE)
-        .find('input').as('searchInput');
+        .find('input')
+        .as('searchInput');
       cy.get(SEARCH_INPUT_CY.AUTOCOMPLETE)
-        .find('chi-dropdown').as('dropdown');
+        .find('chi-dropdown')
+        .as('dropdown');
       cy.get(SEARCH_INPUT_CY.AUTOCOMPLETE)
-        .find('.chi-dropdown__menu-item').as('dropdownMenuItem');
+        .find('.chi-dropdown__menu-item')
+        .as('dropdownMenuItem');
     });
 
     afterEach(() => {
@@ -22,7 +25,11 @@ describe('Search input', () => {
     });
 
     it('Should have mode="autocomplete" attribute to enable autocomplete functionality', () => {
-      cy.get(SEARCH_INPUT_CY.AUTOCOMPLETE).should('have.attr', 'mode', 'autocomplete');
+      cy.get(SEARCH_INPUT_CY.AUTOCOMPLETE).should(
+        'have.attr',
+        'mode',
+        'autocomplete'
+      );
     });
 
     it('Should open the dropdown', () => {
@@ -65,15 +72,15 @@ describe('Search input', () => {
     it('Should be able navigate between the menu items with keyboard arrow up and down', () => {
       cy.get('@searchInput')
         .focus()
-        .then(($el) => {
+        .then($el => {
           cy.get($el)
             .type('{downArrow}{downArrow}{upArrow}')
             .then(() => {
               cy.get('@dropdownMenuItem')
                 .first()
                 .should('have.focus');
-          });
-      })
+            });
+        });
     });
   });
-})
+});
