+++ cypress/e2e/chi-js/accordion.cy.js
