+++ cypress/e2e/chi-js/chi-tabs.cy.js
