+++ cypress/e2e/chi-js/enterprise-global-nav.cy.js
