+++ cypress/e2e/chi-js/range-slider.cy.js
