+++ cypress/e2e/chi-js/sidenav-accordion.cy.js
