+++ cypress/e2e/chi-js/sidenav.cy.js
