+++ cypress/e2e/chi-vue/column-customization.cy.js
