+++ cypress/e2e/chi-vue/expansion-panel.cy.js
