+++ cypress/e2e/chi-vue/search-input.cy.js
