--- cypress/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "baseUrl": "../node_modules",
-    "types": [
-      "cypress"
-    ]
-  },
-  "include": [
-    "**/*.*"
-  ]
-}
