+++ package.json
@@ -36,9 +36,9 @@
     "autoprefixer": "9.4.9",
     "babel-loader": "8.1.0",
     "backstopjs": "4.5.1",
-    "browser-sync": "2.28.3",
+    "browser-sync": "2.29.3",
     "cssnano": "4.1.10",
-    "cypress": "4.12.1",
+    "cypress": "12.17.3",
     "del": "3.0.0",
     "eslint-loader": "2.1.2",
     "gulp": "4.0.0",
@@ -55,7 +55,7 @@
     "gulp-sass": "4.0.2",
     "gulp-sass-lint": "1.4.0",
     "gulp-sri": "0.3.1",
-    "gulp-svg-sprite": "1.5.0",
+    "gulp-svg-sprite": "2.0.3",
     "gulp-wait": "0.0.2",
     "jstransformer-marked": "1.0.3",
     "jstransformer-pug": "0.3.0",
