+++ src/boilerplates/react/chi-react-boilerplate/public/index.html
@@ -24,7 +24,7 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <title>React App</title>
-    <link rel="stylesheet" href="https://assets.ctl.io/chi/5.37.0/chi.css" integrity="sha256-yzGoHKXTuMGTZCEt/JDMTGUFvj0rQxbLLG7+pnLmjCs=" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://assets.ctl.io/chi/5.37.0/chi.css" integrity="sha256-yRzxnZXDbVGA2v0Q9N9HIxRR1VX0aS2atSVOrURKREY=" crossorigin="anonymous">
     <link rel="icon" type="image/svg+xml" href="https://assets.ctl.io/chi/5.37.0/assets/themes/lumen/images/favicon.svg" integrity="sha256-s8cfYbPxhcrLj+FL/OxALhZoDmlCwvX2/K3I0hq1PAs=" crossorigin="anonymous">
     <link rel="alternate icon" href="https://assets.ctl.io/chi/5.37.0/assets/themes/lumen/images/favicon.ico" integrity="sha256-EkKmbH+i/VIQAtUl7NF4bPVaaJZCeBc5xWx8LTcMJp0=" crossorigin="anonymous">
     <script src="https://assets.ctl.io/chi/5.37.0/js/chi.js" integrity="sha256-6QhYmHCoTdqje2hbaXewfi4/GRD7ar8PaJNc/txRYpw=" crossorigin="anonymous"></script>
