+++ src/boilerplates/vue/chi-vue-boilerplate/public/index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <title>Vue Boilerplate</title>
-    <link rel="stylesheet" href="https://assets.ctl.io/chi/5.37.0/chi.css" integrity="sha256-yzGoHKXTuMGTZCEt/JDMTGUFvj0rQxbLLG7+pnLmjCs=" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://assets.ctl.io/chi/5.37.0/chi.css" integrity="sha256-yRzxnZXDbVGA2v0Q9N9HIxRR1VX0aS2atSVOrURKREY=" crossorigin="anonymous">
     <link rel="icon" type="image/svg+xml" href="https://assets.ctl.io/chi/5.37.0/assets/themes/lumen/images/favicon.svg" integrity="sha256-s8cfYbPxhcrLj+FL/OxALhZoDmlCwvX2/K3I0hq1PAs=" crossorigin="anonymous">
     <link rel="alternate icon" href="https://assets.ctl.io/chi/5.37.0/assets/themes/lumen/images/favicon.ico" integrity="sha256-EkKmbH+i/VIQAtUl7NF4bPVaaJZCeBc5xWx8LTcMJp0=" crossorigin="anonymous">
     <script src="https://assets.ctl.io/chi/5.37.0/js/chi.js" integrity="sha256-6QhYmHCoTdqje2hbaXewfi4/GRD7ar8PaJNc/txRYpw=" crossorigin="anonymous"></script>
