+++ src/custom-elements/docs/docs.json
@@ -1,5 +1,5 @@
 {
-  "timestamp": "2023-07-11T11:48:00",
+  "timestamp": "2023-07-19T10:28:17",
   "compiler": {
     "name": "@stencil/core",
     "version": "1.8.9",
